gh-113317: Move Argument Clinic converters to sub-modules
Move the following classes to new sub-modules:

* libclinic.return_converters:

  * CReturnConverterAutoRegister
  * CReturnConverter

* libclinic.int_converters:

  * bool_converter
  * char_converter
  * unsigned_char_converter
  * byte_converter
  * short_converter
  * unsigned_short_converter
  * int_converter
  * unsigned_int_converter
  * long_converter
  * unsigned_long_converter
  * long_long_converter
  * unsigned_long_long_converter
  * Py_ssize_t_converter
  * slice_index_converter
  * size_t_converter

* libclinic.obj_converters:

  * fildes_converter
  * float_converter
  * double_converter
  * Py_complex_converter
  * object_converter
  * buffer
  * rwbuffer
  * robuffer
  * str_converter
  * PyBytesObject_converter
  * PyByteArrayObject_converter
  * unicode_converter
  * Py_UNICODE_converter
  * Py_buffer_converter

* libclinic.misc_converters:

  * defining_class_converter
  * self_converter

Move also Null, NULL and TypeSet to libclinic.utils.

Add create_python_parser_namespace() for PythonParser to pass names
to exec().

Fix --converters command in run_clinic(). Don't use dir(clinic) to
list converters, but use 'converters' and 'return_converters'
dictionaries instead.
