+++ Lib/test/test_clinic.py
@@ -2657,6 +2657,7 @@ class ClinicExternalTest(TestCase):
                 float()
                 int()
                 long()
+                object()
                 Py_ssize_t()
                 size_t()
                 unsigned_int()
