+++ Tools/clinic/libclinic/__init__.py
@@ -32,6 +32,9 @@ from .utils import (
     Sentinels,
     unspecified,
     unknown,
+    TypeSet,
+    Null,
+    NULL,
 )
 
 
@@ -68,6 +71,9 @@ __all__ = [
     "Sentinels",
     "unspecified",
     "unknown",
+    "TypeSet",
+    "Null",
+    "NULL",
 ]
