+++ Tools/clinic/libclinic/misc_converters.py
@@ -1,0 +1,151 @@
+from libclinic.converter import CConverter
+from libclinic.crenderdata import CRenderData, TemplateDict
+from libclinic.function import (
+    Function, Parameter,
+    CALLABLE, METHOD_INIT, GETTER, SETTER,
+    METHOD_NEW, CLASS_METHOD, STATIC_METHOD)
+
+
+
+class defining_class_converter(CConverter):
+    """
+    A special-case converter:
+    this is the default converter used for the defining class.
+    """
+    type = 'PyTypeObject *'
+    format_unit = ''
+    show_in_signature = False
+
+    def converter_init(self, *, type: str | None = None) -> None:
+        self.specified_type = type
+
+    def render(self, parameter: Parameter, data: CRenderData) -> None:
+        self._render_self(parameter, data)
+
+    def set_template_dict(self, template_dict: TemplateDict) -> None:
+        template_dict['defining_class_name'] = self.name
+
+
+def correct_name_for_self(
+        f: Function
+) -> tuple[str, str]:
+    if f.kind in {CALLABLE, METHOD_INIT, GETTER, SETTER}:
+        if f.cls:
+            return "PyObject *", "self"
+        return "PyObject *", "module"
+    if f.kind is STATIC_METHOD:
+        return "void *", "null"
+    if f.kind in (CLASS_METHOD, METHOD_NEW):
+        return "PyTypeObject *", "type"
+    raise AssertionError(f"Unhandled type of function f: {f.kind!r}")
+
+
+class self_converter(CConverter):
+    """
+    A special-case converter:
+    this is the default converter used for "self".
+    """
+    type: str | None = None
+    format_unit = ''
+
+    def converter_init(self, *, type: str | None = None) -> None:
+        self.specified_type = type
+
+    def pre_render(self) -> None:
+        f = self.function
+        default_type, default_name = correct_name_for_self(f)
+        self.signature_name = default_name
+        self.type = self.specified_type or self.type or default_type
+
+        kind = self.function.kind
+
+        if kind is STATIC_METHOD or kind.new_or_init:
+            self.show_in_signature = False
+
+    # tp_new (METHOD_NEW) functions are of type newfunc:
+    #     typedef PyObject *(*newfunc)(PyTypeObject *, PyObject *, PyObject *);
+    #
+    # tp_init (METHOD_INIT) functions are of type initproc:
+    #     typedef int (*initproc)(PyObject *, PyObject *, PyObject *);
+    #
+    # All other functions generated by Argument Clinic are stored in
+    # PyMethodDef structures, in the ml_meth slot, which is of type PyCFunction:
+    #     typedef PyObject *(*PyCFunction)(PyObject *, PyObject *);
+    # However!  We habitually cast these functions to PyCFunction,
+    # since functions that accept keyword arguments don't fit this signature
+    # but are stored there anyway.  So strict type equality isn't important
+    # for these functions.
+    #
+    # So:
+    #
+    # * The name of the first parameter to the impl and the parsing function will always
+    #   be self.name.
+    #
+    # * The type of the first parameter to the impl will always be of self.type.
+    #
+    # * If the function is neither tp_new (METHOD_NEW) nor tp_init (METHOD_INIT):
+    #   * The type of the first parameter to the parsing function is also self.type.
+    #     This means that if you step into the parsing function, your "self" parameter
+    #     is of the correct type, which may make debugging more pleasant.
+    #
+    # * Else if the function is tp_new (METHOD_NEW):
+    #   * The type of the first parameter to the parsing function is "PyTypeObject *",
+    #     so the type signature of the function call is an exact match.
+    #   * If self.type != "PyTypeObject *", we cast the first parameter to self.type
+    #     in the impl call.
+    #
+    # * Else if the function is tp_init (METHOD_INIT):
+    #   * The type of the first parameter to the parsing function is "PyObject *",
+    #     so the type signature of the function call is an exact match.
+    #   * If self.type != "PyObject *", we cast the first parameter to self.type
+    #     in the impl call.
+
+    @property
+    def parser_type(self) -> str:
+        assert self.type is not None
+        if self.function.kind in {METHOD_INIT, METHOD_NEW, STATIC_METHOD, CLASS_METHOD}:
+            tp, _ = correct_name_for_self(self.function)
+            return tp
+        return self.type
+
+    def render(self, parameter: Parameter, data: CRenderData) -> None:
+        """
+        parameter is a clinic.Parameter instance.
+        data is a CRenderData instance.
+        """
+        if self.function.kind is STATIC_METHOD:
+            return
+
+        self._render_self(parameter, data)
+
+        if self.type != self.parser_type:
+            # insert cast to impl_argument[0], aka self.
+            # we know we're in the first slot in all the CRenderData lists,
+            # because we render parameters in order, and self is always first.
+            assert len(data.impl_arguments) == 1
+            assert data.impl_arguments[0] == self.name
+            assert self.type is not None
+            data.impl_arguments[0] = '(' + self.type + ")" + data.impl_arguments[0]
+
+    def set_template_dict(self, template_dict: TemplateDict) -> None:
+        template_dict['self_name'] = self.name
+        template_dict['self_type'] = self.parser_type
+        kind = self.function.kind
+        cls = self.function.cls
+
+        if kind.new_or_init and cls and cls.typedef:
+            if kind is METHOD_NEW:
+                type_check = (
+                    '({0} == base_tp || {0}->tp_init == base_tp->tp_init)'
+                 ).format(self.name)
+            else:
+                type_check = ('(Py_IS_TYPE({0}, base_tp) ||\n        '
+                              ' Py_TYPE({0})->tp_new == base_tp->tp_new)'
+                             ).format(self.name)
+
+            line = f'{type_check} &&\n        '
+            template_dict['self_type_check'] = line
+
+            type_object = cls.type_object
+            type_ptr = f'PyTypeObject *base_tp = {type_object};'
+            template_dict['base_type_ptr'] = type_ptr
