+++ Tools/clinic/libclinic/utils.py
@@ -1,3 +1,4 @@
+import builtins as bltns
 import collections
 import enum
 import hashlib
@@ -82,3 +83,15 @@ class Sentinels(enum.Enum):
 
 unspecified: Final = Sentinels.unspecified
 unknown: Final = Sentinels.unknown
+
+
+TypeSet = set[bltns.type[object]]
+
+
+# This one needs to be a distinct class, unlike the other two
+class Null:
+    def __repr__(self) -> str:
+        return '<Null>'
+
+
+NULL = Null()
